update for tuylocal 2024.11.0
replace 
"Disconnected" to "available" 
 "Connected" to "plugged_in"

Now, when you plug in the car, Home Assistant doesn’t move the toggle to "start charging". Because of this, the code can't switch back to "stop charging" since it’s already in the "stop" state. So, I added a 5-second "start charging" trigger to then switch back to "stop" afterward.